fix: コンテナ・プレゼンテーションパターンで、入力周りの関数の責務を分離
