--- virtual_partner_front/src/components/TextField/index.tsx
@@ -1,2 +1,0 @@
-import { TextFieldView } from './TextFiledView';
-export { TextFieldView as TextField }
