+++ virtual_partner_front/src/features/TextField/index.tsx
@@ -1,0 +1,2 @@
+import { TextFieldContainer } from "./components/TextFieldContainer";
+export { TextFieldContainer as TextField };
